Documents/requirements_win_3.7.5.txt
altgraph==0.17
asgiref==3.3.4
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
comtypes==1.1.14
config==0.5.1
cowsay==5.0
cryptography==3.4.6
cycler==0.10.0
databases==0.4.3
decorator==4.4.2
Django==3.2.4
dnspython==2.1.0
fire==0.4.0
Flask==1.1.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.28
hexdump==3.3
icon-font-to-png==0.4.1
idna==2.10
image==1.5.33
imageio==2.21.3
imageio-ffmpeg==0.4.7
impacket==0.9.22
importlib-metadata==4.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
ldap3==2.9
ldapdomaindump==0.9.3
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.2
moviepy==1.0.3
mysql==0.0.3
mysqlclient==2.0.3
Naked==0.1.31
netaddr==0.8.0
ntlm-auth==1.5.0
numpy==1.21.2
opencv-python==4.6.0.66
palettable==3.3.0
paramiko==2.7.2
pefile==2021.5.24
Pillow==8.4.0
playsound==1.3.0
proglog==0.1.10
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.15.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyminifier==2.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySide2==5.15.2
PySocks==1.7.1
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
rsa==4.9
scipy==1.7.0
shellescape==3.8.1
shiboken2==5.15.2
smbprotocol==1.0.2.dev0
smmap==5.0.0
socks==0
soupsieve==2.3.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
stylecloud==0.5.2
termcolor==2.0.1
tinycss==0.4
tqdm==4.64.0
typing-extensions==3.10.0.0
urllib3==1.26.3
Werkzeug==1.0.1
wordcloud==1.8.2.2
zipp==3.5.0
